pyproject.toml
[tool.poetry]
name = "ochessr"
version = "0.1.0"
description = ""
authors = ["Hugh Cunningham <hugh.e.cunningham@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tensorflow = "^2.3.0"
pillow = "^7.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
